# docs/api/openapi-b.yaml
...
openapi: "3.1.0"
info: 
  title: "ai-shop B 端 BFF"
  version: "0.1.0"
  license: 
    name: "UNLICENSED"
  description: "由 b-app/src/api/endpoints.ts + api/requests.ts + packages/shared/types 自动生成，请勿手改。\n生成命令：cd b-app && npm run gen:api\n\n口径：响应包 {code,msg,data}，分页 {records,total,page,size}，camelCase，单号 xxxNo，时间 xxxAt（UTC 毫秒），枚举大写下划线，金额为最小货币单位整数。"
servers: 
  - url: "http://localhost:8080"
    description: "本地后端"
components: 
  securitySchemes: 
    bearerAuth: 
      type: "http"
      scheme: "bearer"
  schemas: 
...
    CurrencyCode: 
      type: "string"
      enum: 
        - "CNY"
        - "USD"
        - "AED"
...
    TrafficSource: 
      type: "string"
      enum: 
        - "MERCHANT_OWNED"
        - "PLATFORM"
      description: "流量来源。**与 ops-web 的 `TrafficSource` 同名** —— 那边多 INVITE/CHANNEL 两个值（已标 MERGE）"
...
    MerchantTodo: 
      type: "object"
      properties: 
        toShip: 
          type: "number"
          description: "待发货单数（EXPRESS 履约）"
        toDeliver: 
          type: "number"
          description: "待自送单数（商家自送履约）"
        toStock: 
          type: "number"
          description: "待备货单数（自提单已付款，货还没送到自提点）。**按门店算**，这是供货方的活。\n\n与  {@link  toPick }  是同一批单的两头，**两个数不相等**： 买家常常选别家的自提点。`toPick` 按自提点算（我要在点上分多少）， 这一个按门店算（我要送出去多少）。"
        toVerify: 
          type: "number"
          description: "待核销单数（自提到货、买家还没来取）"
        toPick: 
          type: "number"
          description: "待分拣单数（到货后按商品汇总点数）"
        afterSale: 
          type: "number"
          description: "待处理售后单数"
        toReply: 
          type: "number"
          description: "待回复的评价数"
        quotable: 
          type: "number"
          description: "可报价的求团需求数"
      required: 
        - "toShip"
        - "toDeliver"
        - "toStock"
        - "toVerify"
        - "toPick"
        - "afterSale"
        - "toReply"
        - "quotable"
      additionalProperties: false
      description: "工作台待办。**数字即入口** —— 商家打开 App 只想知道「有几件事要我做」"
    MerchantStats: 
      type: "object"
      properties: 
        todayOrders: 
          type: "number"
          description: "今日订单数（自然日，按市场本地时区切分）"
        todayGmvMinor: 
          type: "number"
          description: "今日成交额（最小货币单位）"
        monthOrders: 
          type: "number"
          description: "本月订单数"
        monthGmvMinor: 
          type: "number"
          description: "本月成交额（最小货币单位）"
        currency: 
          "$ref": "#/components/schemas/CurrencyCode"
          description: "统计口径的币种"
        rating: 
          type: "number"
          description: "店铺综合评分，0–5"
        ratingCount: 
          type: "number"
          description: "参与评分的评价条数"
        ownedTrafficRate: 
          type: "number"
          description: "自带客流占比（trafficSource=MERCHANT_OWNED），决定费率档（ADR-004 §6）"
      required: 
        - "todayOrders"
        - "todayGmvMinor"
        - "monthOrders"
        - "monthGmvMinor"
        - "currency"
        - "rating"
        - "ratingCount"
        - "ownedTrafficRate"
      additionalProperties: false
...
    ShareKit: 
      type: "object"
      properties: 
        text: 
          type: "string"
          description: "分享文案，已按当前语言与市场生成"
        posterUrl: 
          type: "string"
          description: "分享海报图 URL"
      required: 
        - "text"
        - "posterUrl"
      additionalProperties: false
      description: "分享素材（B-11.2.7）。文案与海报由服务端按当前语言与市场生成"
...
    MerchantCustomer: 
      type: "object"
      properties: 
        nickname: 
          type: "string"
          description: "客户昵称"
        avatar: 
          type: "string"
          description: "客户头像"
        orderCount: 
          type: "number"
          description: "在本店的累计下单次数"
        totalSpentMinor: 
          type: "number"
          description: "在本店的累计消费额（最小货币单位）"
        lastOrderAt: 
          type: "number"
          description: "最近一次下单时间"
        daysSinceLast: 
          type: "number"
          description: "距上次下单天数"
        silent: 
          type: "boolean"
          description: "沉默客户：曾经常来、最近没来。**这是店主唯一能立刻行动的信号**"
        source: 
          "$ref": "#/components/schemas/TrafficSource"
          description: "客流来源：他是你自己带来的，还是平台分配的"
      required: 
        - "nickname"
        - "avatar"
        - "orderCount"
        - "totalSpentMinor"
        - "lastOrderAt"
        - "daysSinceLast"
        - "silent"
        - "source"
      additionalProperties: false
      description: "商家的客户（B-11.2.8）。\n\n这是「商家自带客流」定位下最该给店主看的东西：**谁在买、谁不来了**。 平台电商给商家看的是流量与转化；小店老板要的是「张阿姨上个月每周都来，这半个月没来」。"
...
paths: 
...
  "/biz/store/share-kit": 
    get: 
      operationId: "mShareKit"
      summary: "分享素材"
      tags: 
        - "store"
      security: 
        - bearerAuth: []
      parameters: 
        - name: "goodsNo"
          in: "query"
          required: false
          schema: 
            type: "string"
            description: "不传即整店分享素材；传了就是单品的"
      responses: 
        "200": 
          description: "OK"
          content: 
            "application/json": 
              schema: 
                type: "object"
                properties: 
                  code: 
                    type: "integer"
                    description: "0 表示成功"
                  msg: 
                    type: "string"
                  data: 
                    "$ref": "#/components/schemas/ShareKit"
                required: 
                  - "code"
                  - "msg"
                  - "data"
  "/biz/dashboard/todo": 
    get: 
      operationId: "mTodo"
      summary: "工作台待办"
      tags: 
        - "dashboard"
      security: 
        - bearerAuth: []
      parameters: []
      responses: 
        "200": 
          description: "OK"
          content: 
            "application/json": 
              schema: 
                type: "object"
                properties: 
                  code: 
                    type: "integer"
                    description: "0 表示成功"
                  msg: 
                    type: "string"
                  data: 
                    "$ref": "#/components/schemas/MerchantTodo"
                required: 
                  - "code"
                  - "msg"
                  - "data"
  "/biz/dashboard/stats": 
    get: 
      operationId: "mStats"
      summary: "经营数据"
      tags: 
        - "dashboard"
      security: 
        - bearerAuth: []
      parameters: []
      responses: 
        "200": 
          description: "OK"
          content: 
            "application/json": 
              schema: 
                type: "object"
                properties: 
                  code: 
                    type: "integer"
                    description: "0 表示成功"
                  msg: 
                    type: "string"
                  data: 
                    "$ref": "#/components/schemas/MerchantStats"
                required: 
                  - "code"
                  - "msg"
                  - "data"
...
  "/biz/customers": 
    get: 
      operationId: "mCustomers"
      summary: "客户与复购"
      tags: 
        - "customers"
      security: 
        - bearerAuth: []
      parameters: []
      responses: 
        "200": 
          description: "OK"
          content: 
            "application/json": 
              schema: 
                type: "object"
                properties: 
                  code: 
                    type: "integer"
                    description: "0 表示成功"
                  msg: 
                    type: "string"
                  data: 
                    type: "array"
                    items: 
                      "$ref": "#/components/schemas/MerchantCustomer"
                required: 
                  - "code"
                  - "msg"
                  - "data"
...